Don't have hosts inherit products from guests
This is correct as far as UNLIMITED_VIRTUALIZATION subscriptions are concerned,
but it is not correct, for example, in the case of a PHYSICAL_ONLY subscriptino
referring to a product that is only installed in VMs.

Instead, change matchability rules so that a subscription can be assigned to
a host even if it does not have a matching product (but one of its VMs does),
and change matching rules so that they don't rely anymore on installation
relationship but only on possible matches.