Distinguish bundled free matches from regular ones in Drools
This makes it possible to limit the OptaPlanner phase to the regular ones only,
reducing the search space dramatically.

This also avoids spending several steps when exchanging a subscription previously
matched to a virtual host, as this required to "detach" it from all virtual guests
first.

Incidentally, this fixes problems we had because the score function had a trap
in those conditions.