Add rule for preferring matches with same centgroup, increase the step count
Motivation for the new rule: We'd like to prefer matches that share same
cent group since it's means more coverage for the same price.

Motivation for increasing step count: Previously we had step count 3K.
On the big datasets (>100K matches), it could happen that matcher is on
a way to discover a better solution (the score is constantly improving
in local search), but it's still terminated after 3000 steps.