DO NOT MERGE: Merge Unlimited virtualization SUMA lifecycle products with their single variant
This is a workaround at the moment and we should get the real merged
products from SCC in the future.

The problem is we differentiate between lifecycle _products_ based on
the policy. E.g. we have 2 different products: Mgmt Lifecycle Product
for Unlimited virtualization and Mgmt Lifecycle Product for Single
systems. Why? In case of SLES or other products we don't make this
difference.